requirements.txt
anthropic>=0.40.0
mcp>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
httpx>=0.26.0
